backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
openai>=1.54.0
pdfplumber==0.10.3
Pillow>=10.2.0
pillow-heif>=0.16.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytz==2023.3

